Revert "add new js files"
