+++ js-structure/js-structure_quiz-app/index.html
@@ -7,7 +7,7 @@
     <link rel="shortcut icon" href="./public/favicon.ico" type="image/x-icon" />
     <title>JS Structure: Quiz-App</title>
     <link rel="stylesheet" href="./styles/styles.css" />
-    <script src="./index.js" type="module"></script>
+    <script src="./index.js" defer></script>
   </head>
   <body>
     <div id="root"></div>
