+++ js-structure/js-structure_quiz-app/index.js
@@ -1,36 +1,209 @@
-/*
-1. refractor each block to different files
-2. add export (+default) to each block
-3. on index, create new variable to call back to on every block we exported ^
-4. add return x on every new block refracted if inside new function
-5. add import xxx on each file for wtv outside elements it mentions
+const cards = [
+  {
+    question:
+      "In the Kingdom Hearts series who provides the english voice for Master Eraqus?",
+    answer: "Mark Hamill",
+    tags: ["kingdomheart", "mastereraqus", "english voices"],
+    isBookmarked: false,
+  },
+  {
+    question:
+      "Daniel Radcliffe became a global star in the film industry due to his performance in which film franchise?",
+    answer: "Harry Potter",
+    tags: ["hewhomustnotbenamed", "danielradcliffe", "abracadabra"],
+    isBookmarked: false,
+  },
+  {
+    question:
+      'In the game series "The Legend of Zelda" what is the name of the main character?',
+    answer: "Link",
+    tags: ["linktothepast", "linkbetweenworlds", "linkbetweenages"],
+    isBookmarked: false,
+  },
+];
 
+function Header() {
+  const header = document.createElement("header");
+  header.classList.add("header");
+  header.innerHTML = /* html */ `
+	  <h1 class="header__title">Quiz-App</h1>
+	`;
 
-*/
+  return header;
+}
 
-// -------------------------- IMPORTS --------------------------
+function handleFormSubmit(event) {
+  event.preventDefault();
 
-import CardsArray from "../utils/cardarray.js";
+  const formData = new FormData(event.target);
+  const data = Object.fromEntries(formData);
 
-import App from "../utils/app.js";
+  const newCard = {
+    question: data.question,
+    answer: data.answer,
+    tags: data.tags.split(","),
+    isBookmarked: false,
+  };
 
-const cardarrayBlock = CardsArray();
+  cards.unshift(newCard);
 
-// -------------------------- HEADER --------------------------
+  // To prevent the need to rerender, we add the new card to the DOM directly
+  const cardList = document.querySelector(".card-list");
+  const newCardElement = Card(newCard);
+  cardList.prepend(newCardElement);
 
-// -------------------------- SUBMIT EVENT - NEW CARD --------------------------
+  event.target.reset();
+  event.target.elements.question.focus();
+}
 
-// -------------------------- FORM --------------------------
+function Form() {
+  const form = document.createElement("form");
+  form.classList.add("form");
+  form.innerHTML = /* html */ `
+		<label for="question">Your question:</label>
+		<textarea
+			id="question"
+			rows="7"
+			name="question"
+			maxlength="150"
+			data-js="input-question"
+		></textarea>
+		<span class="form__character-count">
+			<span data-js="amount-left-question"></span> characters left
+		</span>
+		<label for="answer">Your answer:</label>
+		<textarea
+			id="answer"
+			rows="7"
+			name="answer"
+			maxlength="150"
+			data-js="input-answer"
+		></textarea>
+		<span class="form__character-count">
+			<span data-js="amount-left-answer"></span> characters left
+		</span>
+		<label for="tags">Tags:</label>
+		<input id="tags" type="text" name="tags" />
+		<button class="form__submit-button" type="submit">Submit</button>
+	`;
 
-// -------------------------- BOOKMARKS --------------------------
+  const formFields = form.querySelectorAll('[data-js*="input"]');
+  const counterOutputs = form.querySelectorAll("[data-js*=amount-left]");
 
-// -------------------------- CARDS --------------------------
+  formFields.forEach((formField, index) => {
+    counterOutputs.forEach((output) => {
+      output.innerText = formField.maxLength;
+    });
+    formField.addEventListener("input", () => {
+      const currentAmountLeft = formField.maxLength - formField.value.length;
+      counterOutputs[index].innerText = currentAmountLeft;
+    });
+  });
 
-// -------------------------- CARDLIST --------------------------
+  form.addEventListener("submit", handleFormSubmit);
 
-// -------------------------- APP --------------------------
+  return form;
+}
 
-const appBlock = App();
+function Bookmark(props) {
+  const bookmark = document.createElement("button");
+  bookmark.classList.add("bookmark");
+
+  if (props.active) {
+    bookmark.classList.add("bookmark--active");
+  }
+
+  bookmark.type = "button";
+  bookmark.setAttribute("aria-label", "bookmark");
+  bookmark.innerHTML = /* html */ `
+		<svg
+			class="bookmark__icon"
+			xmlns="http://www.w3.org/2000/svg"
+			viewbox="0 0 24 24"
+		>
+			<path
+				d="M17 3H7c-1.1 0-2 .9-2 2v16l7-3 7 3V5c0-1.1-.9-2-2-2z"
+			/>
+		</svg>
+	`;
+
+  bookmark.addEventListener("click", props.onClick);
+
+  return bookmark;
+}
+
+function Card(props) {
+  const card = document.createElement("article");
+  card.classList.add("card");
+  card.innerHTML = /* html */ `
+		<h2 class="card__question" data-js="question"></h2>
+		<button class="card__button-answer" type="button" data-js="answer-button">
+			Show answer
+		</button>
+		<p class="card__answer" data-js="answer"></p>
+		<ul class="card__tag-list" data-js="tags"></ul>
+		<div class="card__button-bookmark" data-js="bookmark-button-wrapper"></div>
+	`;
+
+  const question = card.querySelector('[data-js="question"]');
+  const answerButton = card.querySelector('[data-js="answer-button"]');
+  const answer = card.querySelector('[data-js="answer"]');
+  const tags = card.querySelector('[data-js="tags"]');
+  const bookmarkButtonWrapper = card.querySelector(
+    '[data-js="bookmark-button-wrapper"]'
+  );
+
+  question.textContent = props.question;
+  answer.textContent = props.answer;
+
+  props.tags.forEach((tag) => {
+    const tagElement = document.createElement("li");
+    tagElement.classList.add("card__tag-list-item");
+    tagElement.textContent = "#" + tag;
+    tags.append(tagElement);
+  });
+
+  function handleBookmarkButtonClick(event) {
+    cards.find((card) => card.question === props.question).isBookmarked =
+      !props.isBookmarked;
+    // To prevent the need to rerender, we can just toggle the class here
+    event.currentTarget.classList.toggle("bookmark--active");
+  }
+
+  const bookmarkButton = Bookmark({
+    active: props.isBookmarked,
+    onClick: handleBookmarkButtonClick,
+  });
+  bookmarkButtonWrapper.append(bookmarkButton);
+
+  function handleAnswerButtonClick() {
+    answer.classList.toggle("card__answer--active");
+  }
+
+  answerButton.addEventListener("click", handleAnswerButtonClick);
+
+  return card;
+}
+
+function CardList() {
+  const cardList = document.createElement("section");
+  cardList.classList.add("card-list");
+
+  cards.forEach((card) => {
+    const cardElement = Card(card);
+    cardList.append(cardElement);
+  });
+
+  return cardList;
+}
+
+function App() {
+  const app = document.createElement("main");
+  app.classList.add("app");
+  app.append(Header(), Form(), CardList());
+
+  return app;
+}
 
 // The render function renders the app to the DOM
 function render() {
