--- js-structure/js-structure_quiz-app/utils/cardlist.js
@@ -1,15 +1,0 @@
-import CardsArray from "./utils/cardarray.js";
-
-export default function CardList() {
-  const cardList = document.createElement("section");
-  cardList.classList.add("card-list");
-
-  cards.forEach((card) => {
-    const cardElement = Card(card);
-    cardList.append(cardElement);
-  });
-
-  return cardList;
-}
-
-const cardarrayBlock = CardsArray ();
