--- js-structure/js-structure_quiz-app/utils/header.js
@@ -1,9 +1,0 @@
-export default function Header() {
-  const header = document.createElement("header");
-  header.classList.add("header");
-  header.innerHTML = /* html */ `
-	  <h1 class="header__title">Quiz-App</h1>
-	`;
-
-  return header;
-}
