--- js-structure/js-structure_quiz-app/utils/submitevent.js
@@ -1,28 +1,0 @@
-import CardsArray from "./utils/cardarray.js";
-import CardList from "./utils/cardlist.js";
-
-export default function handleFormSubmit(event) {
-  event.preventDefault();
-
-  const formData = new FormData(event.target);
-  const data = Object.fromEntries(formData);
-
-  const newCard = {
-    question: data.question,
-    answer: data.answer,
-    tags: data.tags.split(","),
-    isBookmarked: false,
-  };
-
-  cards.unshift(newCard);
-
-  // To prevent the need to rerender, we add the new card to the DOM directly
-  const cardList = document.querySelector(".card-list");
-  const newCardElement = Card(newCard);
-  cardList.prepend(newCardElement);
-
-  event.target.reset();
-  event.target.elements.question.focus();
-}
-const cardArrayBlock = CardsArray();
-const cardListBlock = CardList();
